Split a CIDE::Docker module out
